Extract all the calculators into one AverageCalculator file.
